Write fakeDataUtils test
Init test

Ended write test

Ended write test
